fuzz tc proposal
