+++ monad-consensus/Cargo.toml
@@ -28,3 +28,4 @@ monad-state = { path = "../monad-state" }
 monad-testutil = { path = "../monad-testutil" }
 
 test-case = { workspace = true }
+
