+++ monad-consensus/fuzz/src/lib.rs
@@ -1,0 +1,1 @@
+pub mod fuzz_test_utils;
